changes in admin panel of prerak view and learner view
